Make test more robust to nonfatal condition WSAEWOULDBLOCK
Check for it after calling osaSocketServer::Accept method, and retry
after a brief Sleep up to 10 times before failing.

Perhaps this sort of thing should be worked into the implementation of
osaSocketServer::Accept so that clients do not have to have platform
specific code just to be able to use it reliably on Windows.